Removed some more signed/unsigned compiler warnings.

+++ src-library/MidiEvent.cpp
@@ -57,7 +57,7 @@ MidiEvent::MidiEvent(const MidiEvent& mfevent) {
    seconds = mfevent.seconds;
    eventlink = NULL;
    this->resize(mfevent.size());
-   for (int i=0; i<this->size(); i++) {
+   for (int i=0; i<(int)this->size(); i++) {
       (*this)[i] = mfevent[i];
    }
 }
@@ -104,7 +104,7 @@ MidiEvent& MidiEvent::operator=(MidiEvent& mfevent) {
    seconds = mfevent.seconds;
    eventlink = NULL;
    this->resize(mfevent.size());
-   for (int i=0; i<this->size(); i++) {
+   for (int i=0; i<(int)this->size(); i++) {
       (*this)[i] = mfevent[i];
    }
    return *this;
@@ -117,7 +117,7 @@ MidiEvent& MidiEvent::operator=(MidiMessage& message) {
    }
    clearVariables();
    this->resize(message.size());
-   for (int i=0; i<this->size(); i++) {
+   for (int i=0; i<(int)this->size(); i++) {
       (*this)[i] = message[i];
    }
    return *this;
@@ -127,7 +127,7 @@ MidiEvent& MidiEvent::operator=(MidiMessage& message) {
 MidiEvent& MidiEvent::operator=(vector<uchar>& bytes) {
    clearVariables();
    this->resize(bytes.size());
-   for (int i=0; i<this->size(); i++) {
+   for (int i=0; i<(int)this->size(); i++) {
       (*this)[i] = bytes[i];
    }
    return *this;
